refactor: 상수, 유틸함수 분리
