+++ src/basic/main.basic.js
@@ -1,23 +1,12 @@
+import { DISCOUNT_RATE, prodList } from './constants';
+import { startRandomlyInMs } from './utils';
+
 const store = {
   lastSelectedId: null,
   bonusPts: 0,
   totalAmt: 0,
   itemCnt: 0,
 };
-
-// 할인율
-const DISCOUNT_RATE = {
-  tuesday: 0.1,
-  bulk: 0.25,
-};
-
-const prodList = [
-  { id: 'p1', name: '상품1', price: 10000, quantity: 50, discountRate: 0.1 },
-  { id: 'p2', name: '상품2', price: 20000, quantity: 30, discountRate: 0.15 },
-  { id: 'p3', name: '상품3', price: 30000, quantity: 20, discountRate: 0.2 },
-  { id: 'p4', name: '상품4', price: 15000, quantity: 0, discountRate: 0.05 },
-  { id: 'p5', name: '상품5', price: 25000, quantity: 10, discountRate: 0.25 },
-];
 
 function main() {
   initLayout();
@@ -96,11 +85,6 @@ const startSuggestion = () => {
     suggestionItem.price = Math.round(suggestionItem.price * 0.95);
     updateSelectOptions();
   }
-};
-
-const startRandomlyInMs = (ms) => (callback) => {
-  const startsAt = Math.random(0) * ms;
-  setTimeout(callback, startsAt);
 };
 
 // 변경된 상품 목록으로 select, option 초기화
