+++ src/basic/utils.js
@@ -1,0 +1,4 @@
+export const startRandomlyInMs = (ms) => (callback) => {
+  const startsAt = Math.random(0) * ms;
+  setTimeout(callback, startsAt);
+};
